requirements.txt
aiohttp==3.8.3
aiohttp-jinja2==1.5
faker==15.3.4
torch==1.10.1
numpy==1.22.3
omegaconf==2.1.1
nemo-toolkit==1.6.2
pytorch-lightning==1.5.9
transformers==4.19.1
hydra-core==1.1.1
sacrebleu==1.5.1
braceexpand==0.1.7
webdataset==0.1.62
h5py==3.6.0
matplotlib==3.5.1
rapidfuzz==1.9.1
youtokentome==1.0.6
sacremoses==0.0.49
numba==0.56.0
jieba==0.42.1
opencc==1.1.3
pangu==4.0.6.1
